bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord==1.7.3
discord.py==1.7.3
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-pubsub==2.11.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
multidict==6.0.2
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2

python-dotenv
gitpython
pygithub
flask
gunicorn